[lib][uthread] Change user_addr_t to 32 bit
We only support 32 bit user-space currently.

+++ lib/uthread/include/uthread.h
@@ -36,7 +36,7 @@
 #include <kernel/vm.h>
 #endif
 
-typedef uintptr_t user_addr_t;
+typedef uint32_t user_addr_t;
 
 typedef struct uthread_map
 {
